pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "markdown>=3.7",
    "openai>=1.65.4",
    "py-solc-x>=2.0.3",
    "pyflakes==3.2.0",
    "pygments==2.19.1",
    "rich==13.9.4",
    "weasyprint>=64.1",
    "web3>=7.8.0",
]
